requirements/server-updated.txt
# Core dependencies - Updated and secured
faster-whisper==1.1.5
websockets==13.1
onnxruntime==1.20.1
numba==0.60.0
kaldialign==0.9.1
soundfile==0.12.1
scipy==1.14.1
av==13.1.0
jiwer==3.0.4
evaluate==0.4.3
numpy==2.1.3
openai-whisper==20240930
tokenizers==0.20.3
transformers[torch]==4.46.3
sentencepiece==0.2.0

# OpenVINO backend
librosa==0.10.2
openvino==2024.5.0
openvino-genai==2024.5.0
openvino-tokenizers==2024.5.0
optimum==1.23.3
optimum-intel==1.20.0

# Security & Authentication
pyjwt==2.10.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.18

# Rate limiting & Caching
slowapi==0.1.9
redis==5.2.0
aiocache==0.12.3

# Monitoring & Logging
prometheus-client==0.21.0
structlog==24.4.0
python-json-logger==2.0.7
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0

# API & Validation
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.2
python-dotenv==1.0.1

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0

# Development
black==24.10.0
flake8==7.1.1
mypy==1.13.0
pre-commit==4.0.1